# QuitSmoke/Controllers/Entities/FRSData.cs
public class FRSData
{
    public int age { get; set; }
    public string gender { get; set; }
    public int total_cholesterol { get; set; }
    public int hdl_cholesterol { get; set; }
    public int systolic_blood_pressure { get; set; }
    public bool isTreated { get; set; }
}