# QuitSmoke/Controllers/Entities/UpdatePartner.cs
using System;
using System.Collections;

namespace QuitSmokeWebAPI.Controllers.Entity
{
    public class UpdatePartner
    {
        public string smokerUID { get; set; }
        public string partnerEmail { get; set; }
    }
}