# QuitSmoke/openapi.yaml
swagger: '2.0'
info:
  version: v1
  title: Quit Smoke API
host: quit-smoke-web-api.appspot.com
paths:
  /api/Values:
    get:
      tags:
        - Values
      operationId: ApiValuesGet
      consumes: []
      produces:
        - text/plain
        - application/json
        - text/json
      parameters: []
      responses:
        '200':
          description: Success
          schema:
            uniqueItems: false
            type: array
            items:
              type: string
    post:
      tags:
        - Values
      operationId: ApiValuesPost
      consumes:
        - application/json-patch+json
        - application/json
        - text/json
        - application/*+json
      produces:
        - text/plain
        - application/json
        - text/json
      parameters:
        - name: newUser
          in: body
          required: false
          schema:
            $ref: '#/definitions/AppUser'
      responses:
        '200':
          description: Success
          schema:
            type: string
  '/api/Values/{id}':
    get:
      tags:
        - Values
      operationId: ApiValuesByIdGet
      consumes: []
      produces:
        - text/plain
        - application/json
        - text/json
      parameters:
        - name: id
          in: path
          required: true
          type: integer
          format: int32
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/MyTestEntity'
    put:
      tags:
        - Values
      operationId: ApiValuesByIdPut
      consumes:
        - application/json-patch+json
        - application/json
        - text/json
        - application/*+json
      produces: []
      parameters:
        - name: id
          in: path
          required: true
          type: integer
          format: int32
        - name: value
          in: body
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Success
    delete:
      tags:
        - Values
      operationId: ApiValuesByIdDelete
      consumes: []
      produces: []
      parameters:
        - name: id
          in: path
          required: true
          type: integer
          format: int32
      responses:
        '200':
          description: Success
  /api/Values/login:
    post:
      tags:
        - Values
      operationId: ApiValuesLoginPost
      consumes:
        - application/json-patch+json
        - application/json
        - text/json
        - application/*+json
      produces:
        - text/plain
        - application/json
        - text/json
      parameters:
        - name: user
          in: body
          required: false
          schema:
            $ref: '#/definitions/AppUser'
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/AppUser'
  /api/Values/calculatefrs:
    post:
      tags:
        - Values
      operationId: ApiValuesCalculatefrsPost
      consumes:
        - application/json-patch+json
        - application/json
        - text/json
        - application/*+json
      produces:
        - text/plain
        - application/json
        - text/json
      parameters:
        - name: frsData
          in: body
          required: false
          schema:
            $ref: '#/definitions/FRSData'
      responses:
        '200':
          description: Success
          schema:
            format: int32
            type: integer
  /api/Values/checkEmail:
    post:
      tags:
        - Values
      operationId: ApiValuesCheckEmailPost
      consumes:
        - application/json-patch+json
        - application/json
        - text/json
        - application/*+json
      produces:
        - text/plain
        - application/json
        - text/json
      parameters:
        - name: email
          in: body
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Success
          schema:
            type: boolean
  /api/Values/updatePartner:
    post:
      tags:
        - Values
      operationId: ApiValuesUpdatePartnerPost
      consumes:
        - application/json-patch+json
        - application/json
        - text/json
        - application/*+json
      produces:
        - text/plain
        - application/json
        - text/json
      parameters:
        - name: updatePartner
          in: body
          required: false
          schema:
            $ref: '#/definitions/UpdatePartner'
      responses:
        '200':
          description: Success
          schema:
            type: boolean
definitions:
  AppUser:
    type: object
    properties:
      email:
        type: string
      password:
        type: string
      name:
        type: string
      smoker_indicator:
        type: string
      partner_indicator:
        type: string
      register_date:
        type: string
      suburb:
        type: string
      city:
        type: string
      plan_id:
        type: string
      partner_id:
        type: string
      point:
        format: int32
        type: integer
      uid:
        type: string
      age:
        type: string
      gender:
        type: string
  MyTestEntity:
    type: object
    properties:
      name:
        type: string
      unit:
        $ref: '#/definitions/Unit'
  Unit:
    type: object
    properties:
      fiT5046:
        type: string
      fiT5140:
        type: string
  FRSData:
    type: object
    properties:
      age:
        format: int32
        type: integer
      gender:
        type: string
      total_cholesterol:
        format: int32
        type: integer
      hdl_cholesterol:
        format: int32
        type: integer
      systolic_blood_pressure:
        format: int32
        type: integer
      isTreated:
        type: boolean
  UpdatePartner:
    type: object
    properties:
      smokerUID:
        type: string
      partnerEmail:
        type: string